Revert "Hotfixes and multiple votings for GoC"
