+++ .vscode/launch.json
@@ -1,13 +1,22 @@
 {
+  // Use IntelliSense to learn about possible attributes.
+  // Hover to view descriptions of existing attributes.
+  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
   "version": "0.2.0",
   "configurations": [
-      {
-          "type": "node",
-          "request": "attach",
-          "name": "Node: Nodemon",
-          "processId": "${command:PickProcess}",
-          "restart": true,
-          "protocol": "inspector",
-      },
+    {
+      "name": "Launch via NPM",
+      "type": "node",
+      "request": "launch",
+      "cwd": "${workspaceRoot}",
+      "runtimeExecutable": "npm",
+      "runtimeArgs": [
+          "run-script", "debug"
+      ],
+      "port": 2222,
+      "skipFiles": [
+        "<node_internals>/**"
+      ],
+    }
   ]
 }
