+++ backend/.env.example
@@ -17,7 +17,6 @@ SENDGRID_GENERIC_TEMPLATE=
 SENDGRID_CONTACT_TEMPLATE=
 SENDGRID_CONTACT_MAIL=hello@hackjunction.com
 MONGODB_URI=
-TEST_MONGODB_URI=
 FRONTEND_URL=
 DEVTOOLS_ENABLED=
 ADMIN_TOKEN=
