+++ backend/modules/project_score/model.js
@@ -27,10 +27,6 @@ const ProjectScoreSchema = new mongoose.Schema(
             type: String,
             default: null,
         },
-        scoreGiver: {
-            type: String,
-            required: false,
-        },
         score: {
             type: Number,
             required: false,
@@ -51,7 +47,7 @@ const ProjectScoreSchema = new mongoose.Schema(
     },
     { toJSON: { virtuals: true } },
 )
-/* ProjectScoreSchema.index(
+ProjectScoreSchema.index(
     {
         project: 1,
         event: 1,
@@ -61,7 +57,7 @@ const ProjectScoreSchema = new mongoose.Schema(
     {
         unique: true,
     },
-) */
+)
 
 ProjectScoreSchema.set('timestamps', true)
