+++ backend/package.json
@@ -7,13 +7,12 @@
         "dev": "nodemon index.js",
         "start": "node index.js | npm run pipe-logs",
         "pipe-logs": "pino-cloudwatch --aws_access_key_id=$AWS_ACCESS_KEY_ID --aws_secret_access_key=$AWS_SECRET_ACCESS_KEY --aws_region=$AWS_REGION --group=$AWS_CLOUDWATCH_LOG_GROUP",
-        "test": "NODE_ENV=test jest --verbose --runInBand --detectOpenHandles",
+        "test": "echo \"Error: no test specified\" && exit 1",
         "lint": "eslint ./common && eslint ./modules && eslint ./utils",
         "lint-fix": "eslint ./common --fix && eslint ./modules --fix && eslint ./utils --fix",
         "depcheck": "depcheck .",
         "eslint-check": "eslint --print-config path/to/main.js | eslint-config-prettier-check",
-        "precommit": "echo \"Backend: running pre-commit checks\"",
-        "debug": "nodemon --inspect index.js"
+        "precommit": "echo \"Backend: running pre-commit checks\""
     },
     "author": "",
     "license": "ISC",
@@ -112,12 +111,9 @@
         "eslint-config-prettier": "^6.11.0",
         "eslint-plugin-import": "^2.20.2",
         "eslint-plugin-prettier": "^3.1.3",
-        "jest": "^26.6.3",
         "kill-port": "^1.6.0",
         "nodemon": "^2.0.4",
-        "prettier": "^2.0.5",
-        "superagent": "^6.1.0",
-        "supertest": "^6.1.3"
+        "prettier": "^2.0.5"
     },
     "engines": {
         "node": "12.3.1",
