+++ frontend/src/components/projects/ProjectsGrid/index.js
@@ -7,41 +7,6 @@ import { EventHelpers } from '@hackjunction/shared'
 import ProjectsGridItem from '../ProjectsGridItem'
 
 import ProjectScoresService from 'services/projectScores'
-import project from 'pages/_dashboard/slug/project'
-
-// Helper function to calculate project scores average.
-const averageScores = ({ avg, n }, scores) => {
-    return {
-        avg: (scores.score + n * avg) / (n + 1),
-        n: n + 1,
-    }
-}
-/*
-if project has been multiple scores calculate the average of the scores, and add it as score.
-Note using 'score' key for multiple score average is a quite dirty hack to sort them.
-*/
-//TODO: Change 'score' - key to average or something
-const projectScoresAverage = (inital, project) => {
-    if ('scores' in project) {
-        const initialVals = { avg: 0, n: 0 }
-        const averageProjectScores = project.scores.reduce(
-            averageScores,
-            initialVals,
-        ).avg
-        const result = [...inital, { ...project, score: averageProjectScores }]
-        console.log('result', result)
-        return result
-    } else {
-        return [...inital, project]
-    }
-}
-//add projects into same list
-const addAverageToProjects = projects => {
-    const inital = []
-    const final = projects.reduce(projectScoresAverage, inital)
-    //console.log('all', all)
-    return final
-}
 
 const ProjectsGrid = ({
     projects,
@@ -53,9 +18,8 @@ const ProjectsGrid = ({
     token = '',
 }) => {
     const isOngoingEvent = EventHelpers.isEventOngoing(event, moment)
-    const isVotingOpen = EventHelpers.isVotingOpen(event, moment)
     const [sorted, setSorted] = useState(projects)
-    console.log('event>>', event)
+
     async function fetchData() {
         const nprojects = await Promise.all(
             projects.map(async project => {
@@ -65,25 +29,13 @@ const ProjectsGrid = ({
                     project._id,
                 )
                     .then(score => {
-                        //TODO: refactor this one
-                        if (!isVotingOpen) {
-                            if (score[0]) {
-                                return { ...{ scores: [...score] }, ...project }
-                            }
-                            return Object.assign(
-                                { score: 0, message: 'Not rated' },
-                                project,
-                            )
-                        } else {
-                            return Object.assign(
-                                {
-                                    score: 0,
-                                    message:
-                                        'Review ongoing. Results will be shown after review period is over',
-                                },
-                                project,
-                            )
+                        if (score[0]) {
+                            return Object.assign(score[0], project)
                         }
+                        return Object.assign(
+                            { score: 0, message: 'Not rated' },
+                            project,
+                        )
                     })
                     .catch(e => {
                         console.log(e)
@@ -94,10 +46,7 @@ const ProjectsGrid = ({
                     })
             }),
         )
-        const sorted = addAverageToProjects(nprojects)
-        //const result = sortBy(nprojects, p => -p['score'])
-        //console.log(sorted)
-        setSorted(sortBy(sorted, p => -p['score']))
+        setSorted((sortBy(nprojects, p => -p['score']): nprojects))
     }
 
     useEffect(() => {
@@ -110,7 +59,7 @@ const ProjectsGrid = ({
             )
         }
     }, [projects])
-    //    console.log(sorted)
+    console.log(sorted)
     return (
         <Grid
             container
